Update script and gitignore
Crowdin API v1 requires login={account} and account-key={key} to work.
Not sure what the previous api format was from but doesn't work that way

Set up to read account name from .crowdin.account
Added the extra file to gitignore to not have it in the repo.

TODO: document this process.